requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.18.0
sqlalchemy==2.0
psycopg2-binary==2.9
python-dotenv==1.0
passlib==1.7
python-jose==3.3
pydantic-settings==2.8.1
alembic==1.13
pytest==8.0
httpx==0.27
pyrebase4==4.8.0
pytz==2025.2
firebase-admin==6.2.0
pydantic[email]